refactor out tpu_quic_connection
